moved descriptor views and forms from admin to descriptor
