+++ app/templates/admin/index.html
@@ -33,10 +33,6 @@
                                     description='Create a new user account', icon='add user icon') }}
                 {{ dashboard_option('Invite New User', 'admin.invite_user',
                                     description='Invites a new user to create their own account', icon='add user icon') }}
-                {{ dashboard_option('Add New Descriptor', 'admin.new_descriptor',
-                                    description='Create a new descriptor for resources.', icon='add user icon') }}
-                {{ dashboard_option('Existing Descriptors', 'admin.existing_descriptors',
-                                    description='View and edit existing descriptors.', icon='add user icon') }}
             </div>
         </div>
     </div>
