+++ app/templates/descriptor/descriptors.html
@@ -3,10 +3,6 @@
 {% block content %}
     <div class="ui stackable grid container">
         <div class="sixteen wide tablet twelve wide computer centered column">
-            <a class="ui basic compact button" href="{{ url_for('admin.index') }}">
-                <i class="caret left icon"></i>
-                Back to dashboard
-            </a>
             <h2 class="ui header">
                 Descriptors
                 <div class="sub header">
@@ -28,7 +24,7 @@
                 </div>
                 <div class="ui right search item">
                     <div class="ui transparent icon input">
-                        <input id="search-descriptors" type="text" placeholder="Search users…">
+                        <input id="search-descriptors" type="text" placeholder="Search descriptors…">
                         <i class="search icon"></i>
                     </div>
                 </div>
@@ -47,7 +43,7 @@
                     </thead>
                     <tbody>
                     {% for d in descriptors | sort(attribute='name') %}
-                        <tr onclick="window.location.href = '{{ url_for('admin.edit_descriptor', desc_id=d.id) }}';">
+                        <tr onclick="window.location.href = '{{ url_for('descriptor.edit_descriptor', desc_id=d.id) }}';">
 
                             <td>{{ d.name }}</td>
                             {% if not d.values %}
