--- app/templates/descriptor/index.html
@@ -1,7 +1,0 @@
-{% extends 'layouts/base.html' %}
-
-{% block content %}
-  <div class="ui text container">
-    This is a page for descriptors.
-  </div>
-{% endblock %}
