+++ app/templates/descriptor/new_descriptor.html
@@ -7,10 +7,6 @@
 {% block content %}
     <div class="ui stackable centered grid container">
         <div class="twelve wide column">
-            <a class="ui basic compact button" href="{{ url_for('admin.index') }}">
-                <i class="caret left icon"></i>
-                Back to dashboard
-            </a>
             <h2 class="ui header">
                 Add New Descriptor
                 <div class="sub header">Create a new option or text descriptor</div>
